remove use of `pytest_plugins`
