--- backend/tests/fixtures/__init__.py
@@ -1,1 +1,0 @@
-from tests.fixtures.database import *  # noqa: F403
