--- backend/tests/fixtures/database.py
@@ -1,37 +1,0 @@
-import pytest
-from testcontainers.postgres import PostgresContainer
-import sqlalchemy as sa
-import database
-from sqlalchemy.orm import Session
-
-
-@pytest.fixture(scope="session")
-def postgres_container():
-    with PostgresContainer("postgres:13") as postgres:
-        yield postgres
-
-
-@pytest.fixture
-def engine(postgres_container: PostgresContainer, monkeypatch):
-    test_engine = sa.create_engine(
-        sa.URL.create(
-            "postgresql",
-            username=postgres_container.username,
-            password=postgres_container.password,
-            host=postgres_container.get_container_host_ip(),
-            port=postgres_container.get_exposed_port(postgres_container.port),
-            database=postgres_container.dbname,
-        )
-    )
-
-    monkeypatch.setattr(database, "engine", test_engine)
-
-    database.create_db_and_tables()
-
-    return test_engine
-
-
-@pytest.fixture
-def session(engine):
-    with Session(database.engine) as session:
-        yield session
